Fix checks and rehookup breaking changes warning
