+++ src/pkg/packager/components.go
@@ -35,11 +35,11 @@ func (p *Packager) getValidComponents() []types.ZarfComponent {
 			key = component.Name
 		} else {
 			// Otherwise, add the component name to the choice group list for later validation
-			choiceComponents = p.appendIfNotExists(choiceComponents, component.Name)
+			choiceComponents = helpers.AppendIfNotExists(choiceComponents, component.Name)
 		}
 
 		// Preserve component order
-		orderedKeys = p.appendIfNotExists(orderedKeys, key)
+		orderedKeys = helpers.AppendIfNotExists(orderedKeys, key)
 
 		// Append the component to the list of components in the group
 		componentGroups[key] = append(componentGroups[key], component)
@@ -206,15 +206,6 @@ func (p *Packager) confirmChoiceGroup(componentGroup []types.ZarfComponent) type
 	}
 
 	return componentGroup[chosen]
-}
-
-func (p *Packager) appendIfNotExists(slice []string, item string) []string {
-	for _, s := range slice {
-		if s == item {
-			return slice
-		}
-	}
-	return append(slice, item)
 }
 
 func (p *Packager) requiresCluster(component types.ZarfComponent) bool {
